requirements.txt
Pillow
requests
pandas
flask
logging
werkzeug
pystray